// lib/ts/types.ts
export type FeatureBaseProps = {
    /*
     * recipeId.
     */
    recipeId: string;

    /*
     * Children element
     */
    children?: JSX.Element;

    /*
     * History provided by react-router
     */
    history?: any;

    /*
     * Nested Features
     * This is used to avoid reinitialising feature wrapper for nested features.
     */
    isEmbedded?: boolean;
};